아이템 점수 획득 작업 2차
